Regroup layer voltage calculation
